feat: add mobile add button
