+++ linkbrary/src/pages/Folder.js
@@ -40,6 +40,7 @@ function Folder() {
         <Contents items={data} />
       ) : (
         <NoLink>저장된 링크가 없습니다.</NoLink>
+        <AddFolderBtn>폴더 추가 +</AddFolderBtn>
       )}
       <Footer />
     </>
@@ -55,6 +56,31 @@ const NoLink = styled.p`
   font-weight: 400;
   line-height: 24px; /* 150% */
   margin-top: 40px;
+`;
+
+const AddFolderBtn = styled.button`
+  border: none;
+  border-radius: 20px;
+  border: 1px solid var(--Linkbrary-white, #fff);
+  background: var(--Linkbrary-primary-color, #6d6afe);
+  position: sticky;
+  left: 40%;
+  bottom: 101px;
+  padding: 8px 24px;
+  display: none;
+
+  color: var(--Linkbrary-gray10, #e7effb);
+  text-align: center;
+  font-family: Pretendard;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  letter-spacing: -0.3px;
+
+  @media (max-width: 774px) {
+    display: block;
+  }
 `;
 
 export default Folder;
